Remove useless alignment from MpegEncContext.intra_scantable
The first member of struct ScanTable is a simple pointer, extra alignment
of which serves no purpose.  The alignment specifier was added along with
some Altivec optimisations also adding a 16-byte-aligned array at the end
of struct ScanTable.  Presumably the redundant, outer alignment was added
by mistake.  The inner one is clearly sufficient.

git-svn-id: file:///var/local/repositories/ffmpeg/trunk@19724 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b