Implement a new registration system for filters.
Create a new static array containing pointers to the AVFilter
definitions, so that the non-constant next filter in the AVFilter
struct is not anymore required and the AVFilter definitions may be
stored in shareable memory.

Also change the signature for avfilter_register(), make it return an
int since it may fail if there is not enough space in the static array
for the registered filters.


git-svn-id: file:///var/local/repositories/ffmpeg/trunk@20605 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b