Add _XOPEN_SOURCE=600 to CPPFLAGS on NetBSD.
This make NetBSD compile out of the box at the cost of a hack, but a
locally contained one that is preferable to the one used in the wild.
Currently the NetBSD FATE box adds -D_NETBSD_SOURCE to CPPFLAGS to work
around broken system headers. Since it is unlikely for NetBSD to fix their
headers, it is better to use a standard flag instead of a system-specific one.

As a sideeffect, this allows getting rid of a local _NETBSD_SOURCE definition.


git-svn-id: file:///var/local/repositories/ffmpeg/trunk@22486 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b